+++ tests/index.test.ts
@@ -1,0 +1,2 @@
+import './openAlexMapper.test'
+import './zoteroResponse.test'